fix tests and remove switch in new layout manager
