+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.scheduler/virtual_scrolling.integration.tests.js
@@ -657,10 +657,8 @@ module('Virtual scrolling integration', () => {
 
                         const { instance } = this.scheduler;
 
-                        const layoutManager = instance.getLayoutManager();
-                        const settings = layoutManager._positionMap[0][0];
-
-                        assert.equal(settings.groupIndex, 0, 'groupIndex is correct');
+                        const items = instance.getAppointmentsInstance().option('items');
+                        assert.equal(items[0].groupIndex, 0, 'groupIndex is correct');
                     });
                 });
 
@@ -1194,23 +1192,22 @@ module('Virtual scrolling integration', () => {
                                 promise,
                                 () => {
 
-                                    const layoutManager = instance.getLayoutManager();
-                                    const settings = layoutManager._positionMap[0][0];
+                                    const items = instance.getAppointmentsInstance().option('items');
 
                                     assert.equal(
-                                        settings.groupIndex,
+                                        items[0].groupIndex,
                                         0,
                                         `group index is correct when scrolled to ${scrollY}`
                                     );
 
                                     const startViewDate = viewDataProvider.findGroupCellStartDate(
-                                        settings.groupIndex,
-                                        settings.info.appointment.startDate,
-                                        settings.info.appointment.endDate
+                                        items[0].groupIndex,
+                                        items[0].info.appointment.startDate,
+                                        items[0].info.appointment.endDate
                                     );
 
                                     assert.deepEqual(
-                                        settings.info.appointment.startDate,
+                                        items[0].info.appointment.startDate,
                                         startViewDate,
                                         'start date is correct'
                                     );
@@ -1255,10 +1252,8 @@ module('Virtual scrolling integration', () => {
 
                     const { instance } = this.scheduler;
 
-                    const layoutManager = instance.getLayoutManager();
-                    const settings = layoutManager._positionMap[0][0];
-
-                    assert.equal(settings.groupIndex, 0, 'groupIndex is correct');
+                    const items = instance.getAppointmentsInstance().option('items');
+                    assert.equal(items[0].groupIndex, 0, 'groupIndex is correct');
                 });
 
                 test(`Grouped appointment should contains correct groupIndex if "${viewName}" view has horizontal group orientation`, async function(assert) {
@@ -1291,10 +1286,9 @@ module('Virtual scrolling integration', () => {
 
                     const { instance } = this.scheduler;
 
-                    const layoutManager = instance.getLayoutManager();
-                    const settings = layoutManager._positionMap[0][0];
+                    const items = instance.getAppointmentsInstance().option('items');
 
-                    assert.equal(settings.groupIndex, 1, 'groupIndex is correct');
+                    assert.equal(items[0].groupIndex, 1, 'groupIndex is correct');
                 });
 
                 test(`Grouped allDay appointment should contains correct groupIndex if "${viewName}" view has vertical group orientation`, async function(assert) {
@@ -1323,11 +1317,9 @@ module('Virtual scrolling integration', () => {
                     });
 
                     const { instance } = this.scheduler;
+                    const items = instance.getAppointmentsInstance().option('items');
 
-                    const layoutManager = instance.getLayoutManager();
-                    const settings = layoutManager._positionMap[0][0];
-
-                    assert.equal(settings.groupIndex, 0, 'groupIndex is correct');
+                    assert.equal(items[0].groupIndex, 0, 'groupIndex is correct');
                 });
             });
         });
@@ -1412,18 +1404,14 @@ module('Virtual scrolling integration', () => {
                             assert,
                             promise,
                             () => {
-                                instance.getLayoutManager().filteredItems.forEach((dataItem, index) => {
-                                    const layoutManager = instance.getLayoutManager();
-                                    const appointmentRenderingStrategy = layoutManager.getRenderingStrategyInstance();
-                                    const settings = appointmentRenderingStrategy.generateAppointmentSettings(dataItem);
-                                    const {
-                                        groupIndex,
-                                        topPositions
-                                    } = expectedSettings[index];
-                                    assert.equal(settings.length, topPositions.length, 'Settings amount is correct');
+                                expectedSettings.forEach(({
+                                    groupIndex,
+                                    topPositions
+                                }) => {
+                                    const items = instance.getLayoutManager().filteredItems.filter((item) => item.groupIndex === groupIndex).sort();
+                                    assert.equal(items.length, topPositions.length, 'Settings amount is correct');
                                     topPositions.forEach((top, index) => {
-                                        assert.equal(settings[index].groupIndex, groupIndex, `Appointment groupIndex "${groupIndex}" is correct for offsetY: ${offset.y}`);
-                                        assert.equal(settings[index].top, top, `Appointment top position "${top}" is correct for offsetY: ${offset.y}`);
+                                        assert.equal(items[index].top, top, `Appointment top position "${top}" is correct for offsetY: ${offset.y}`);
                                     });
                                 });
                             },
